requirements.txt
# need to install
torch==2.2.2
numpy==1.26.4
matplotlib==3.8.4
torchvision==0.17.2
tqdm==4.66.2
scikit-learn==1.4.2
opencv-python==4.9.0.80
pillow==10.3.0
mediapipe==0.10.14
PyQt5==5.15.10

## additional stuff if needed

# test
seaborn==0.13.2

# model display
torchsummary==1.5.1